Apply suggestions from code review

+++ esphome/components/epaper_spi/epaper_waveshare_b.h
@@ -5,7 +5,7 @@ namespace esphome::epaper_spi {
 
 /**
  * Waveshare (B) series BWR e-paper displays using SSD1680-compatible controllers.
- * Waveshare use 0=red, 1=no-red, the inverse of EPaperWeAct3C
+ * Waveshare uses 0=red, 1=no-red, the inverse of EPaperWeAct3C
  */
 class EpaperWaveshareB : public EPaperWeAct3C {
  public:
@@ -13,7 +13,7 @@ class EpaperWaveshareB : public EPaperWeAct3C {
 
  protected:
   bool reset() override;
-  uint8_t transform_red_byte(uint8_t byte) const override { return ~byte; }
+  uint8_t transform_red_byte(uint8_t byte) const override { return static_cast<uint8_t>(~byte); }
 };
 
 }  // namespace esphome::epaper_spi
